Introduce the concept of in use entries.
Entries that are in use are immune from eviction. This concept is needed
to protect the bits that are currently being referenced by a UIImage/CGImageRef
from being overwritten.

Note also that inUse is a refcounted concept since multiple FICImageTableEntry's
can be allocated against the same entityUUID/bits in the backing file.

A side effect of this change is that the imageFormat's maximumCount is no
longer a hard limit since it is possible that all maximumCount entries could
be in use.